fix tests to work with auth

+++ sensor-ms/app/api/users_api.py
@@ -2,7 +2,7 @@
 # Date created: 5.12.2024
 
 # REST FastAPI dependencies.
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Depends
 from fastapi.responses import PlainTextResponse
 
 # OAuth2 authentication dependencies.
@@ -23,12 +23,13 @@ router = APIRouter(
 )
 
 @router.post("/create-user", status_code=200, response_class=PlainTextResponse)
-async def create_user(user_schema : schema.UserCreate):
+async def create_user(user_schema : schema.UserCreate, token : str = Depends(token_bearer) ):
     """
     This endpoint allows creating a new user in the system.
 
     Args:
         user_schema (UserCreate): The details of the user to create.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during user creation.
@@ -44,12 +45,13 @@ async def create_user(user_schema : schema.UserCreate):
     return result
 
 @router.get("/{username}", response_model=User)
-async def get_user(username: str):
+async def get_user(username: str, token : str = Depends(token_bearer) ):
     """
     This endpoint fetches the details of a user from the system.
 
     Args:
         username (str): The username of the user to retrieve.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during user retrieval.
@@ -58,6 +60,9 @@ async def get_user(username: str):
         User: The retrieved user details.
     """
 
+    if not await validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.get_user(username)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -65,12 +70,13 @@ async def get_user(username: str):
     return result
 
 @router.delete("/{username}", status_code=200, response_class=PlainTextResponse)
-async def delete_user(username: str):
+async def delete_user(username: str, token : str = Depends(token_bearer) ):
     """
     This endpoint removes a user from the system by their username.
 
     Args:
         username (str): The username of the user to delete.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during user deletion.
@@ -79,6 +85,9 @@ async def delete_user(username: str):
         PlainTextResponse: The username if the user is deleted successfully.
     """
 
+    if not await validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
+
     result = await utils.delete_user(username)
     if isinstance(result, Err):
         raise HTTPException(status_code=result.code, detail=result.message)
@@ -86,12 +95,13 @@ async def delete_user(username: str):
     return result
 
 @router.put("/{username}/delete-sensors", status_code=200, response_class=PlainTextResponse)
-async def delete_sensors(username: str):
+async def delete_sensors(username: str, token : str = Depends(token_bearer) ):
     """
     This endpoint clears all storages in a user's account.
 
     Args:
         username (str): The username of the user whose sensors will be deleted.
+        token (str): Access token generated at login time.
 
     Raises:
         HTTPException: If an error occurs during storages clearing.
@@ -99,6 +109,9 @@ async def delete_sensors(username: str):
     Returns:
         PlainTextResponse: The username if the user sensors are deleted successfully.
     """
+
+    if not await validate_token_with_username(username, token):
+        raise HTTPException(status_code=401, detail="Token username missmatch.")
 
     result = await utils.delete_sensors(username)
     if isinstance(result, Err):
